Recognizer: Use higher distance threshold for double sensor gestures
The reasoning behind this is as follows:

When detecting single sensor gestures, half of the gesture string already matches the single sensor candidates, because it consists solely of '_' letters.
In contrast, when detecting double sensor gestures, it is more likely for there to be mismatches because there is an additional set of inputs to compare.