Injector: Add lager_injector to the repo
Lager_injector is an example lager client implementation.

It is a program that uses liblager_connect to register a series of
gestures and listen for detected gesture messages.

Upon receiving a detected gesture message, it matches it to a
corresponding series of XTestFakeKeyEvent() commands in order to
inject key events via the X window manager.

This allows it to perform tasks such as opening the Chrome web
browser, opening and closing new tabs, loading websites, refreshing,
and zooming in and out.