Recognizer: Add gesture messaging. Add CLI arguments for gesture drawing and file use.
Lager_recognizer now uses liblager_connect to create a gesture
subscription queue. It then spawns a thread that constantly
monitors the queue and updates a local list of subscriptions.

When an input gesture matches a subscription, it sends a message
to the message queue that corresponds to the process that
registered the gesture.

Lager_recognizes now takes CLI arguments for toggling gesture
drawing and for choosing between a static gesture database
and the use of the gesture subscription queue.