Common: Change LaGeR coordinate <-> letter mapping scheme
Old mapping was based on vertical slices of a virtual rhombicuboctahedron,
which led to repeating the north and south pole letters repeatedly on each
slice.

New mapping is based on horizontal slices of a virtual rhombicuboctahedron,
with the north and south poles being their the first and last slice, and the
middle slices having no repeated letters.

The new mapping is also more intuitive to visualize.