Add support for the Leap Motion controller
Add support for the Leap Motion controller and use the pinch
gesture as a form of button to start and finish gestures.

Basic LaGeR conversion is working with the left hand by itself, but
work remains to support both hands concurrently.

The dependencies file has been updated with the instructions for
installing the necessary components.