Initial version of LaGeR machine learning classifier and tools
lager_expander: Expands LaGeR strings by duplicating movement characters and
padding with zeroes until a target length is reached. This is important because
all training data must have the same length (number of features).

lager_file_to_numbers: Converts all LaGeR strings in a file into a numerical
representation where "a" is 1, "b" is 2, ..., "z" is 26. This is needed so the
data can be processed numerically by a machine learning classifier.

lager_ml_utils: Miscellaneous utility functions used by other files in this
directory to manipulate LaGeR strings for machine learning purposes.

ml_recognizer: Prepares a LaGeR string and feeds it to a machine learning
classifier to recognize the corresponding gesture.

ml_trainer: Reads a dataset consisting of LaGeR strings in numerical form
prefixed with their corresponding label, and uses it to train a linear or deep
neural network classifier.