Tweak string expansion and DNN parameters
Strings were being expanded by image resizing via bilinear sampling and padded
with constant values (analogous to an image being expanded smoothly with slowly
vanishing edges). This was causing the insertion of multiple values that were
different from that of the original string.

Now, strings are expanded by image resizing via nearest-neighbor and padded with
edge values (analogous to an image being expanded in a simple way with sharp edges).
This means that the original values are more likely to be preserved.