Lager ML: Support recognition of gestures with two sensors
Adapt the numerical representation of LaGeR strings so that the CSV
is converted to 2-dimensional arrays for each movement, where each
dimension is corresponds to one sensor.

Adjust the ML trainer and recognizer accordingly, as well as the common
functions for expanding strings and converting them between letters and
numbers.

Lower the target number of features (movements) to 224, as 384 was more
than needed (the longest gesture so far, OpenCNN, uses less than 200).