Lager ML: Tweak DNN topology and hyperparameters to improve generalization
Increase complexity to two layers with 100 neurons each, which seems
to generalize better and result in fewer errors when categorizing
inputs from imprecise sensors such as the Xbox controller.

It seems that previous commits that simplified the topology went too far
and instead of improving generalization they limited the model so that it
sometimes it failed to categorize input gestures correctly unless they were
very accurate.

Lower starting learning rate and add decay while increasing the number
of epochs to improve chances of finding a good solution to the model
training.

So far the only con is that the recognition time increased from 5-10 ms
15-20 ms, but that is still very fast.