Lager Calibrator: Create utility to calibrate sensor scale (WIP)
Different sensors can be comfortably used at different scales, but OSVR
normalizes all input to a scale of -1.0 to 1.0.

This utility will determine the normal range of comfortable use for a given
sensor and use it to set the scale factor used by liblager_convert to make
sure that the resulting strings have a similar length regardless of the
input device.

[Work In Progress]